+++ Frends.JSON.QuerySingle/Frends.JSON.QuerySingle/QuerySingle.cs
@@ -1,17 +1,20 @@
 ﻿using Frends.JSON.QuerySingle.Definitions;
+using Frends.JSON.QuerySingle.Helpers;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using System;
 using System.ComponentModel;
 using System.IO;
 using System.Reflection;
 using System.Runtime.Loader;
+using System.Threading;
 
 namespace Frends.JSON.QuerySingle;
 
 /// <summary>
 /// JSON Task.
 /// </summary>
-public class JSON
+public static class JSON
 {
     /// Mem cleanup.
     static JSON()
@@ -28,16 +31,25 @@ public class JSON
     /// </summary>
     /// <param name="input">Input parameters.</param>
     /// <param name="options">Optional parameters.</param>
-    /// <returns>Object { bool Success, dynamic Data }</returns>
-    public static Result QuerySingle([PropertyTab] Input input, [PropertyTab] Options options)
+    /// <param name="cancellationToken">Cancellation token.</param>
+    /// <returns>Object { bool Success, dynamic Data, Error Error }</returns>
+    public static Result QuerySingle([PropertyTab] Input input, [PropertyTab] Options options, CancellationToken cancellationToken)
     {
-        JToken jToken = GetJTokenFromInput(input.Json);
-        JToken result = jToken.SelectToken(input.Query, options.ErrorWhenNotMatched);
+        try
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            JToken jToken = GetJTokenFromInput(input.Json);
+            JToken result = jToken.SelectToken(input.Query, options.ErrorWhenNotMatched);
 
-        if (result == null && options.ErrorWhenNotMatched)
-            throw new JsonException($"No matches found for query '{input.Query}'.");
+            if (result == null && options.ErrorWhenNotMatched)
+                throw new JsonException($"No matches found for query '{input.Query}'.");
 
-        return new Result(true, result);
+            return new Result(true, result);
+        }
+        catch (Exception ex)
+        {
+            return ex.Handle(options);
+        }
     }
 
     private static object GetJTokenFromInput(dynamic json)